(maint) Fix flush/update
When flush was implemented for iis_site, the way it determined whether to update the resource was changed to look at  instead of . This broke the way iis_application_pool relied on flush. To fix this, the call to exists? was put back and the iis_site  method was updated to move the property sets to .