(MODULES-4370) Add detection of the PowerShell version
Previously the module could not determine what scripts to execute as the
PowerShell version was unknown.  This commit adds a PowerShell version class and
method on the base IIS Provider to detect and cache the major PowerShell version.
This will be used in later methods to appropriately execute the correct scripts.