(MODULES-4370) Remove typeinformation generation in JSON
Previously the type information for an object was being inserted into the JSON
documents.  While this does not break any existing functionality, future changes
to the module would have to deal with the extra information.  This commit sets
the NoTypeInformation to false and thereby stops all type information being
created.