(MODULES-4434) IIS Feature management
This adds a type and provider to manage Windows feature installation on
Windows 2008 and up. It only allows IIS features to be installed.

Note this code uses an array of the latest IIS features available to
install, but does not keep track of which subset is available in a given
IIS distribution. We could have kept track but since there are only a
handful of added features from 7.5 to 8.5, it was thought it would be
easier to have one array to check rather than keep a seperate list per
IIS version. In short, we defer to the tooling to tell us what feature is
present in which IIS version and only keep track of the larger list.