(MODULES-4463) Fix Get-Website
Previously the cmdlet `Get-Website -Name` was used to retrieve particular
IIS sites however due to a bug in the Server 2008 version of this cmdlet it
returns all sites.  This commit changes to using the Get-Item and Test-Path
method to retrieve site information, which is more performant that simply
filtering all sites.