(MODULES-4463) Fix creating new sites when no other sites exist
Previously, creating a new website would fail on Windows 2008 due to a bug where
it requires at least one site to exist so that it can calculate the next site
ID number.  This commit changes the creation script so that it checks if any
other site exists.  If not it forces the SiteID to 1 (The default for IIS) or
it leaves it for IIS to determine the Site ID.