(maint) Do not print Windows 2008 powershell support to debug
Previously, when a PowerShell command was sent to debug output it contained all
of the PowerShell code to support the ConvertTo-JSON command however this meant
the actual command that was issued was lost in a sea of messages.  This commit
emits the debug message before all of the PowerShell support code is added.