(MODULES-4818) Fix iis_feature spec
This renames the iis_feature spec file so that it's actually picked up
by the test runs. This also reduces the amount of information
Get-WindowsFeature returns so that there isn't a huge amount of data
being passed back from powershell to puppet.

It also fixes the iis_feature detection so that it correctly returns
true of false if the feature is an IIS feature or not.