(maint) Remove redundant ordering in examples
Prior to this, the README contained example usage that used `require`
and relationship arrows to order directories and acl resources. These
relationships do not need to be explicitly defined as Puppet and the acl
resource automatically make those relationshsips.

This commit removes the redundancy so that the example code is more
concise and simple.