(MODULES-5344) iis_virtual_directory handle slashes in name
Prior to this commit, iis_virtual_directory was not idempotent when the name
attribute contained slashes.

iis_virtual_directory uses the PowerShell New-WebVirtualDirectory command.
New-VirtualDirectory uses Name to set Path, and Path may contain slashes.

With this commit, iis_virtual_directory is idempotent when name contains
backward slashes. Due to the behaviour of Get-Item and Get-WebVirtualDirectory,
name may not contain forward slashes, as documented in the associated ticket.