(MODULES-5429) Fix iis_site idempotency
This commit fixes the start and stop methods on the iis_site
provider by adding error catching to the IIS cmdlet invocations.
It now properly fails if Puppet is unable to start or stop a website.

Previously, when issuing a start or stop command that failed, the
error would be passed to STDOUT and not reconized as an error, so
Puppet would report the state had changed and return success. When
run again, Puppet would again miss the error and report the state
has changed, causing a never ending loop of report notifications.

Unfortunately the error returned by IIS is generic, so the reason
for not being able to start or stop the site is unknowable to Puppet
during the run. It is up to the user to figure out the conflict.