(MODULES-5465) convert version confine from fact to feature
Prior to this commit, if IIS is installed during a run via iis_feature,
other iis_* types fail with 'could not find a suitable provider' errors.
This is caused by confine testing a fact.

This confine tests a feature, which can (see PUP-5985) be re-evaluated,
and updates logproperties.ps1.erb to use PowerShell instead of the fact.