(MODULES-5264) accept UNC paths
Prior to this commit, the IIS module:

Included regular expressions that failed to a match a UNC path.
Attempted to use File.exists? to test what could be a UNC path.
Failed to create or update a virtual directory with a UNC path.

Also, this issue interfered with development and testing:

(Misspelled and) failed to set the iis_virtual_directory :application attribute.