(MODULES-5529) Add authenticationinfo to iis_site
This commit adds the ability to set authenticationinfo to iis_site by
adding a new parameter to iis_site and relevant code to the iis_site
provider. Acceptance test for code is added as well

This commit also modifies self.prefetch to merge
`provider.authenticationinfo` and
`resources[site]['authenticationinfo']`. Both hashes have the correct
information,  but the one that comes from the user may have a subset of
all that can be specified. If we don't merge, then Puppet thinks there
was a change that needs to be corrected.