(MODULES-6062) Fix iis_virtual_directory idempotency
This commit fixes the iis_virtual_directory destroy method to provide
the required paramters to the Remove-WebVirtualDirectory command so that
it successfully removes a virtual directory.

It also uses `@property_hash` instead of `@resource` inside the destroy
method, so that the flush method works properly after removing a virtual
directory.

Lastly, it updates the spec acceptance test to properly create and
destroy virtual directories instead of failing silently.