(MODULES-6144) Add limits to iis_site type and provider
Prior to this commit there was no way to set the limits configurations
for an IIS site using this module. All sites would instead inherit the
default settings for connection timeout, max bandwidth, and maximum
number of connections.

This commit adds site limits to the iis_site type and provider. It
also updates unit and acceptance tests for the new capability.