Create site with alternate port
(MODULES-6385) If a user specifies a port in the binding information,
this change will use the supplied port from the binding information hash
to create the website. This capability can be used to avoid a conflict
on Port 80 if there is already a site (such as Default Web Site) running
on that port. This capability is also require to avoid run time errors
if your manifest specifies that one site should run on Port 80 and a
second site should run on an alternate port.