(MODULES-5465) Add puppet feature suitability
This commit changes the confine for all providers to use a puppet
feature instead of a puppet fact. This allows the provider to install
IIS and then use one of the providers to configure it in the same Puppet
run. A feature is used instead of a fact because a feature can be
re-evaluated during the Puppet run and a fact cannot.

This commit also extracts the IIS detection logic out to a class in
order to isolate the parts that decide what IIS versions we support and
how to determine existance of IIS on the target node. Appropriate spec
tests are added for the class.

Automated acceptance tests were not added for this, as the
spec_helper_acceptance installs IIS before any tests are run. Since this
change is to enable installing and configuring IIS in the same run, we
would have to re-write all of the tests in order to be able to add one
test for this. This has been manually validated instead.