(MODULES-7173) Cannot Start IIS Site with Port 443
This change fixese a bug where a site bound to port 443 in the binding
information hash could not be started. Before version 4.3.0 all sites
were created on Port 80, and only later in the configuration process
would they be given a port binding to port 443 and the https protocol
enabled on the site. After 4.3.0, the port binding was passed to the
New-Website cmdlet by default as a way to avoid conflicts on port 80.
Unfortunately, the https protocol was not enabled by default if the port
binding was 443. When the site was started, as it always is by default,
IIS would throw an error because a site bound to port 443 is not allowed
to start if the HTTPS protocol is not enabled.

This change detects when the HTTPS protocol has been specified as
enabled, or when the user has specified port 443 as the port binding. In
reality checking for both is probably not needed as they are mutual
requirements, but it costs little to do the check, so both are checked.

This change also incorporates additional automated testing to ensure
that future releases do not break HTTPS sites.

This change also modifies the test for creating an SSL enabled site. The
test now does three puppet runs to ensure that the site is in its final
configured state and then tests the properties of the site. Only if
there are still changes being reported by puppet on the third run is
there a real idempotency bug we care about at the moment. MODULES-5561
is tracking an idempotency issue that is very similar, and when that
ticket is signed off we will most likely be able to return this test to
running using the normal it_behaves_like 'an idempotent resource'
helper.