(MAINT) Fix Readme Example
Prior to this change, the first example in the readme me would always
fail. The example failed to either delete or disable the default website
'Default Web Site' that is created when installing the IIS feature.
Another approach might have been to bind the 'minimal' site to another
port, but that could have made the example more in depth into IIS
concepts than needed for the very first example.