(MODULES-5493) Manage Two Same Name Appications
This commit fixes a bug in the this module uses the composite name key
feature. Prior to this change the `sitename` property was specified as
a part of the composite key. Unfortunately, properties are not valid
components of a compsite key and the directive was ignored. This meant
that in truth there was only attribute considered for the naming key,
and thus for uniqueness, which was the `applicationname` parameter.

This change removes the composite key functionality and relies on the
provider to be able to parse out the sitename from the application name
as needed. This change allows a user to specify both the sitename and
the applicationname properties in the title of resource in a way that
maintians uniqueness for catalog compilation so that the module can
manage two applications with the same name under two different web sites
on the same IIS instance. Previously this was not possible due to the
broken namevar functionality.

This change also means that the direction of the slashes used in the
title of an iis_application resource, or any of its parameters no longer
matters, however, for backward compatibility, running the resource
command will still use a Windows style back slash to seperate the
sitename from the application name in the returned title of the
resource, and forward slashes as the path seperator from the remainder
of the path to a nested application folder.