(MODULES-8346) Case Sensitive Paths
This change overrides the property_insync method in the custom type
`Path` type in lib\puppet_x\puppetlabs\iis\property\path.rb. The new
implementation ensures that values for is and should are compared in a
case insensitive manner. Values are not munged, so the proper case
provided by the user is still stored when the object is created, but
this avoids erroneous changes being logged if someone changes the value
in the manifest and no change should actually be made.

Prior to this change, if someone simply changes the case of a path in
a manifest, a corrective change would be logged when the user expectation
would have been an idempotent run with no changes logged because Windows
file paths are not case sensitive.