(MODULES-6061) Fix virtual directory update
This change fixes the module's ability to update the physical path that
a virtual directory points to. This fix is in two parts.

First, the line that attempted to update the sitename a virtual
directory belongs to had to be removed. The site name for a virtual
directory is not a property of the virtual directory when looked at from
the point of view of the xml configuration file. Instead the virtual
directory is a child node of the site. This means that
`Set-ItemProperty` cannot be used to move a virtual directory from one
site to another. Worse, since that property does not actually exist, it
will always result in an error.

Second, since `Set-ItemProperty` in this case is a helper function for
xml manipulation, it ended up being case sensitive. The casing of the
property name `physicalPath` was accidentally spelled `physicalpath`.
Unfortunately, the way this cmdlet works in this context, the missed
property name did not result in an error, and Puppet had no way of
knowing the property didn't match, and was led to believe the change had
been made successfully.