(MODULES-6080) Make virtualdir removal idempotent
Prior to this commit attempts to remove virtual directories
would fail if there were any subfolders or files in them -
functionally, always. This failure is silent as a confirmation
prompt was always rising and never being resolved. Unfortunately,
there is no way to forcibly ignore this prompt from the command-
line - no force option and the confirm flag fails to suppress it.

This commit replaces the call to `Remove-WebVirtualDirectory` with
a call to `Remove-Item` in the IIS provider hierarchy. This is
somewhat more fragile and a break from existing patterns. However,
it is the only way we could do this cleanly, given the limitations
of the PowerShell commands.

This is otherwise not a change of behavior.