(MODULES-7820) Fix "No Managed Code" in app pool
This commit updates the `iis_application_pool` type for the
`:managed_runtime_version` to accept an empty string, `''`,
to specify "No Managed Code" - prior to this commit the
definition was `'""'` which actually errors when specified
in PowerShell.

This commit further updates the documentation to be more
explicit about this and updates the acceptance tests to
verify this functionality.