(MODULES-7700) Ensure app pool state is idempotent
Prior to this commit the call to start or stop an app pool
was made regardless of whether it is currently stopped or
started. On older versions of Windows this would sometimes
cause an error alerting that the app pool was already in
a stopped state.

This commit ensures the command runs if and only if the
app pool state needs to change. This prevents the error
from surfacing.