(MAINT) Rename localized idempotent_apply to fix CI
Prior to this commit the acceptance tests for this module referenced a local
implementation of the idempotent_apply method, clobbering the same method made
available via Litmus. At some point, the clobber stopped working, causing the
test suite to fail immediately with argument errors.

This commit renames the local method to be iis_idempotent_apply and updates all
references to idempotent_apply in the acceptance tests. It does *not* replace
the localized implementation with the litmus one, as this will require a more
extensive refactor of the test suite, to be done at a later time.

This change primarily exists to ensure CI is functional in the short term, which
itself enables the refactor work.