(MODULES-10988) fix require_relative being not relative enough
In some cases running this outside the default puppet agent config this
can cause issues if the libdir does not end in `lib`.

This fix does away with the extraneous level of relativity and thus avoids
the problem. This is the only place in the module where this pattern was
used.