[MODULES-11188] Additional utility changes
Prior to this commit, there was an attempt to fix an issue where puppet
could not handle a trailing '/' at the end of a physicalpath. While it
stopped users from creating resources with trailing '/', it did not fix
existing IIS entities that already had the trailing '/'.

This commit aims to complement the previous code by removing trailing
'/' in physicalpaths for existing resources.