[MODULES-11188] Fixing unit tests
Prior to this commit there were some unit tests that would fail due to
the additional checks inserted in the latest fix. Some of the previous
code would also not work as intended and lint was throwing warnings at
some formatting choices made.

This commit aims to provide a number of small fixes to deal with both,
lint warnings and small errors in previous code that would make unit
tests fail.