[MODULES-11188] Updating README
Prior to this commit several fixes were implemented to deal with the
problem described in the JIRA ticket MODULES-11188. As a result of those
changes, the behaviour of the module regarding physicalpaths was
slighly altered.

This commit aims to update the README file to inform users about the
previously existing issue and the measures that the module will take to
deal with it.