(MODULES-11188) Fix physicalPath on apps and sites
Prior to this commit it looks like we'd chosen the incorrect property
to pull back from the iis config. PhysicalPath != physicalPath and the
former was behaving differently between the app and site resources.

This commit standardises on physicalPath which returns the correct value
for both app and site resources. For clarity, physicalPath returns a
string which we can directly manipulate if required.